requirements.txt
# Document Processing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.3

# Text Processing & NLP
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.35.0

# Change Detection
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Vector DB & Embeddings
chromadb==0.4.15
faiss-cpu==1.7.4
numpy==1.24.3

# QA & LLM Integration
langchain==0.0.330
groq>=0.1.0
# OR for open-source alternatives:
# torch==2.1.0
# accelerate==0.24.1

# Data Processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.4

# Web & API
requests==2.31.0
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7
pyyaml==6.0.1